[add] - board entity 제거 및 board enum 타입 분리, addFunctions 신규 생성 및 명명

+++ src/main/java/com/mlfrog/domain/Board/BoardId.java
@@ -13,8 +13,8 @@ public class BoardId implements Serializable {
 
     private final long value;
 
-    static BoardId of(long value) {
-        return BoardId.of(value);
+    static BoardId of(int value) {
+        return BoardId.of((long) value);
     }
 
     @Override
